Expire an auction after it is paid for via buy it now, or reverse auction.
But there is a question.  How do we know if two people clicked buy it now at the same time and both are checking out.  Do we have to do a check somewhere to make sure they are the first person to actually finish checking out.  Also what if they click buy it now, then wait two hours to finish checking out?